Move wsl utility functions to the utilities file
