+++ vscode-dotnet-runtime-library/src/Acquisition/IDistroDotnetSDKProvider.ts
@@ -319,13 +319,13 @@ If you would like to contribute to the list of supported distros, please visit: 
 
     protected myDistroStrings(stringKey: string): string
     {
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+
         return this.distroJson[this.distroVersion.distro][stringKey];
     }
 
     protected myDistroCommands(commandKey: string): CommandExecutorCommand[]
     {
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+
         return this.distroJson[this.distroVersion.distro][commandKey] as CommandExecutorCommand[];
     }
 
@@ -333,11 +333,11 @@ If you would like to contribute to the list of supported distros, please visit: 
     {
         const validCommands: string[] = [];
 
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+
         const baseCommands = (Object.values(this.distroJson[this.distroVersion.distro])
             .filter((x: any) => x && Array.isArray(x) && ((x[0] as CommandExecutorCommand).commandParts))).flat();
 
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+
         let preInstallCommands = this.myVersionDetails()[this.preinstallCommandKey] as CommandExecutorCommand[];
         if (!preInstallCommands)
         {
@@ -367,15 +367,15 @@ If you would like to contribute to the list of supported distros, please visit: 
     {
         let allPackages: string[] = [];
         // Remove this when https://github.com/typescript-eslint/typescript-eslint/issues/2728 is done
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+
         const distroPackages = this.distroJson[this.distroVersion.distro][this.distroPackagesKey];
         for (const packageSet of distroPackages)
         {
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+
             allPackages = allPackages.concat(packageSet[this.sdkKey]);
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+
             allPackages = allPackages.concat(packageSet[this.runtimeKey])
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+
             allPackages = allPackages.concat(packageSet[this.aspNetKey])
         }
         return allPackages;
