+++ vscode-dotnet-runtime-library/src/Acquisition/LinuxVersionResolver.ts
@@ -13,7 +13,6 @@ import
     DotnetCustomLinuxInstallExistsError,
     DotnetInstallLinuxChecks,
     DotnetUpgradedEvent,
-    DotnetWSLCheckEvent,
     EventBasedError,
     EventCancellationError
 } from '../EventStream/EventStreamEvents';
@@ -23,6 +22,7 @@ import { FileUtilities } from '../Utils/FileUtilities';
 import { ICommandExecutor } from '../Utils/ICommandExecutor';
 import { getInstallFromContext } from '../Utils/InstallIdUtilities';
 import { IUtilityContext } from '../Utils/IUtilityContext';
+import { getRunningDistro, microsoftSupportedDistroIds } from '../Utils/TypescriptUtilities';
 import { DebianDistroSDKProvider } from './DebianDistroSDKProvider';
 import { DotnetInstallMode } from './DotnetInstallMode';
 import { GenericDistroSDKProvider } from './GenericDistroSDKProvider';
@@ -98,124 +98,6 @@ Follow the instructions here to download the .NET SDK: https://learn.microsoft.c
 Or, install Red Hat Enterprise Linux 8.0 or Red Hat Enterprise Linux 9.0 from https://access.redhat.com/downloads/;`
     protected acquireCtx: IDotnetAcquireContext | null | undefined;
 
-    // All distros that we officially support for this tool. If a distro is not in this list, it can still have community member support.
-    public static readonly microsoftSupportedDistroIds = [RED_HAT_DISTRO_INFO_KEY, UBUNTU_DISTRO_INFO_KEY];
-
-    /**
-     * Checks if the system is running under WSL.
-     * Checks env vars first, then falls back to /proc/version.
-     * @param eventStream Optional event stream for diagnostic logging.
-     */
-    public static async isWSL(eventStream?: IEventStream): Promise<boolean>
-    {
-        eventStream?.post(new DotnetWSLCheckEvent(`Checking if system is WSL. OS: ${os.platform()}`));
-
-        if (os.platform() !== 'linux')
-        {
-            return false;
-        }
-
-        if (process.env.WSL_DISTRO_NAME || process.env.WSLENV)
-        {
-            return true;
-        }
-
-        try
-        {
-            const procVersion = await new FileUtilities().read('/proc/version');
-            return procVersion.toLowerCase().includes('microsoft');
-        }
-        catch
-        {
-            return false;
-        }
-    }
-
-    /**
-     * Detects the Linux distro and version from /etc/os-release.
-     * @param eventStream Optional event stream for diagnostic logging.
-     * @returns The distro name and version, or null if it can't be determined.
-     */
-    public static async getRunningDistro(eventStream?: IEventStream): Promise<DistroVersionPair | null>
-    {
-        if (os.platform() !== 'linux')
-        {
-            return null;
-        }
-
-        const mainOSDeclarationFile = `/etc/os-release`;
-        // Some distros may not include the os-release file specified by system d: https://0pointer.de/blog/projects/os-release
-        // This is a recommended fallback: https://man7.org/linux/man-pages/man5/os-release.5.html
-        const backupOSDeclarationFile = `/usr/lib/os-release`;
-        const fileUtils = new FileUtilities();
-        const osDeclarationFile = await fileUtils.exists(mainOSDeclarationFile) ? mainOSDeclarationFile : backupOSDeclarationFile;
-
-        try
-        {
-            const osInfo = (await fileUtils.read(osDeclarationFile)).split('\n');
-            const infoWithQuotesRemoved = osInfo.map(x => x.replace('"', ''));
-            const infoWithSeparatedKeyValues = infoWithQuotesRemoved.map(x => x.split('='));
-            const keyValueMap = Object.fromEntries(infoWithSeparatedKeyValues.map(x => [x[0], x[1]]));
-
-            const distroName: string = keyValueMap.NAME?.replace('"', '') ?? '';
-            const distroVersion: string = keyValueMap.VERSION_ID?.replace('"', '') ?? '';
-
-            if (distroName === '' || distroVersion === '')
-            {
-                return null;
-            }
-
-            return { distro: distroName, version: distroVersion };
-        }
-        catch
-        {
-            return null;
-        }
-    }
-
-    /**
-     * Checks if the given distro (or the current running distro) is in microsoftSupportedDistroIds.
-     * Returns false if the distro can't be determined.
-     * @param distro Optional distro to check. If not provided, detects the current distro.
-     * @param eventStream Optional event stream for diagnostic logging.
-     */
-    public static async isDistroSupported(distro?: DistroVersionPair | null, eventStream?: IEventStream): Promise<boolean>
-    {
-        const resolvedDistro = distro ?? await LinuxVersionResolver.getRunningDistro(eventStream);
-        if (!resolvedDistro || resolvedDistro.distro === '')
-        {
-            return false;
-        }
-        return LinuxVersionResolver.microsoftSupportedDistroIds.includes(resolvedDistro.distro);
-    }
-
-    /**
-     * Checks for WSL or non-Microsoft-supported Linux distro.
-     * Returns { isUnsupported: true, reason } if on WSL or a community/unsupported distro.
-     * Note: the extension itself can still install on community distros (e.g. Debian via DebianDistroSDKProvider).
-     * This method is intended for LM tools that should not attempt community-support installs.
-     * @param eventStream Optional event stream for diagnostic logging.
-     */
-    public static async checkForUnsupportedLinux(eventStream?: IEventStream): Promise<{ isUnsupported: boolean; reason?: string }>
-    {
-        if (os.platform() !== 'linux')
-        {
-            return { isUnsupported: false };
-        }
-
-        if (await LinuxVersionResolver.isWSL(eventStream))
-        {
-            return { isUnsupported: true, reason: 'WSL' };
-        }
-
-        if (!await LinuxVersionResolver.isDistroSupported(undefined, eventStream))
-        {
-            return { isUnsupported: true, reason: 'Linux Distro' };
-        }
-
-        return { isUnsupported: false };
-    }
-
     constructor(private readonly workerContext: IAcquisitionWorkerContext, private readonly utilityContext: IUtilityContext,
         executor: ICommandExecutor | null = null, distroProvider: IDistroDotnetSDKProvider | null = null)
     {
@@ -241,7 +123,7 @@ Or, install Red Hat Enterprise Linux 8.0 or Red Hat Enterprise Linux 9.0 from ht
             return this.distro;
         }
 
-        const result = await LinuxVersionResolver.getRunningDistro(this.workerContext.eventStream);
+        const result = await getRunningDistro(this.workerContext.eventStream);
 
         if (!result || result.distro === '' || result.version === '')
         {
@@ -285,7 +167,7 @@ Or, install Red Hat Enterprise Linux 8.0 or Red Hat Enterprise Linux 9.0 from ht
         }
         else
         {
-            if (!LinuxVersionResolver.microsoftSupportedDistroIds.includes(this.distro.distro))
+            if (!microsoftSupportedDistroIds.includes(this.distro.distro))
             {
                 // UX: Could eventually add a 'Go away' button via the callback:
                 this.utilityContext.ui.showInformationMessage(`Automated SDK installation for the distro ${this.distro.distro} is not officially supported, except for community implemented and Microsoft approved support.
