+++ vscode-dotnet-runtime-library/src/Utils/CommandExecutor.ts
@@ -49,7 +49,6 @@ import { getInstallFromContext } from './InstallIdUtilities';
 
 import { SUDO_LOCK_PING_DURATION_MS } from '../Acquisition/CacheTimeConstants';
 import { IAcquisitionWorkerContext } from '../Acquisition/IAcquisitionWorkerContext';
-import { LinuxVersionResolver } from '../Acquisition/LinuxVersionResolver';
 import { RUN_UNDER_SUDO_LOCK } from '../Acquisition/StringConstants';
 import { IEventStream } from '../EventStream/EventStream';
 import { IWindowDisplayWorker } from '../EventStream/IWindowDisplayWorker';
@@ -61,7 +60,7 @@ import { ICommandExecutor } from './ICommandExecutor';
 import { IFileUtilities } from './IFileUtilities';
 import { IUtilityContext } from './IUtilityContext';
 import { LockUsedByThisInstanceSingleton } from './LockUsedByThisInstanceSingleton';
-import { executeWithLock, loopWithTimeoutOnCond, minimizeEnvironment } from './TypescriptUtilities';
+import { executeWithLock, isRunningUnderWSL, loopWithTimeoutOnCond, minimizeEnvironment } from './TypescriptUtilities';
 
 export class CommandExecutor extends ICommandExecutor
 {
@@ -118,7 +117,7 @@ export class CommandExecutor extends ICommandExecutor
             }
         }
 
-        if (await LinuxVersionResolver.isWSL(this.context?.eventStream))
+        if (await isRunningUnderWSL(this.context?.eventStream))
         {
             // For WSL, vscode/sudo-prompt does not work.
             // This is because it relies on pkexec or a GUI app to popup and request sudo privilege.
@@ -434,7 +433,7 @@ ${stderr}`));
         const fullCommandString = `${command.commandRoot} ${command.commandParts.join(' ')}`;
         let useCache = false;
         // Remove this when https://github.com/typescript-eslint/typescript-eslint/issues/2728 is done
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+
         if (options)
         {
             // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
